requirements.txt
google-generativeai
requests
pydantic
google-adk
python-dotenv